search fixet

+++ API/controllers/foodController.js
@@ -7,37 +7,42 @@ export const getFoods = async (req, res) => {
         let filter = {}
 
         // sets filter to only get approved foods or ones made by user if user is not admin
-        if(req.user.admin == false){
-            filter = {$or: [
-                { godkendt: true }, 
-                { user: req.user.id }
-            ]}
+        if (req.user.admin == false) {
+            filter = {
+                $or: [
+                    { godkendt: true },
+                    { user: req.user.id }
+                ]
+            }
         }
 
         // gets all foods with filter
         const foods = await Food.find(filter);
 
-        res.status(200).json({foods: foods})
+        res.status(200).json({ foods: foods })
     } catch (error) {
         // return status 500 error if failed
         console.log(error)
-        res.status(500).json({msg: "unable to get foods"})
+        res.status(500).json({ msg: "unable to get foods" })
     }
 }
 
 // exporting get method - gets specific food through id
 export const getFood = async (req, res) => {
     try {
-        const id = req.params.id
-        Food.findById(id)
-        .then((foods) => {
-            // returns status 200 and food that matches id
-            res.status(200).json({food: foods})
-        })
+        // finds food by id
+        const food = await Food.findById(req.params.id)
+
+        // checks if user is admin or user who created the food
+        if (!(food.user == req.user.id || req.user.admin)) {
+            return res.status(401).json({ msg: "no access" })
+        }
+
+        // returns the food
+        res.status(200).json({ food })
     } catch (error) {
-        // logs and returns status 500 error if failed or no foods found
-        console.log(error)
-        res.status(500).json({msg: "unable to get food"})
+        // returns status 500 error if failed or no foods found
+        res.status(500).json({ msg: "unable to get food" })
     }
 };
 
@@ -46,36 +51,43 @@ export const search = async (req, res) => {
     try {
         // creates search term from the request query
         const searchTerm = req.query.searchTerm
-        
+        console.log(searchTerm)
+
         // creates regex for search
         const searchRegex = new RegExp(searchTerm, "i")
-        
-        // specifies criterias to search for
-        await Food.find({
-            $or : [
-                {barcode: searchRegex},
-                {name: searchRegex},
-                {calories: searchRegex},
-                {carbonhydrates: searchRegex},
-                {protein: searchRegex},
-                {fat: searchRegex}
+
+        let filter = {
+            $or: [
+                { barcode: searchRegex },
+                { name: searchRegex }
             ]
-        })
-        .then((foods) => {
-            if(foods.length){
-                // logs and returns the foods that match
-                console.log(foods)
-                res.status(200).json({foods: foods})
-            }
-            else{
-                // returns nothing if no foods match
-                res.status(200).json({foods: [], msg: "no foods found"})
-            }
-        })
+        }
+
+        // sets filter to only get approved foods or ones made by user if user is not admin
+        if (req.user.admin == false) {
+            filter = {
+                $and: [
+                    { user: req.user.id },
+                    {
+                        $or: [
+                            { barcode: searchRegex },
+                            { name: searchRegex }
+                        ]
+                    }
+                ]
+            };
+        }
+
+        console.log(filter)
+
+        // searches for foods with the filter
+        const foods = await Food.find(filter)
+
+        // returns the foods
+        res.status(200).json({ foods })
     } catch (error) {
-        // logs and returns status 500 if error
-        console.log(error)
-        res.status(500).json({msg: "unable to get food"})
+        // returns status 500 if error
+        res.status(500).json({ msg: "unable to get food" })
     }
 };
 
@@ -86,26 +98,22 @@ export const createFood = async (req, res) => {
         req.body.user = req.user.id
 
         // makes godkendt false if user is not admin
-        if(req.user.admin == false){
+        if (req.user.admin == false) {
             req.body.godkendt = false
         }
-        
-        // creates a food from food body
-        const food = new Food(req.body)
 
         // saves the new food to database
-        await food.save()
-        .then((savedFoods) => {
-            // logs and returns the created food
-            console.log(savedFoods)
-            res.status(201).json({msg: 'food saved', food})
-        })
+        await new Food(req.body).save()
+            .then((food) => {
+                // logs and returns the created food
+                res.status(201).json({ msg: 'food saved', food })
+            })
     } catch (error) {
         // logs and returns status 500 if error -> food not created
         console.log(error)
-        res.status(500).json({msg: 'unable to save new food'})
+        res.status(500).json({ msg: 'unable to save new food' })
     }
-    
+
 }
 
 // exporting delete method - deletes a food through id
@@ -114,18 +122,30 @@ export const deleteFood = async (req, res) => {
         // gets id from params
         const id = req.params.id
 
-        // deletes the food from database
+        // finds food
+        const food = Food.findById(id)
+
+        // throws error if food is not found
+        if (!food) throw Error("Food not found");
+
+        // 401 response if user is not admin or user who created the food
+        if (!(req.user.admin || req.user.id == food.user)) {
+            return res.status(401).json({ message: 'No access' });
+        }
+
         await Food.findByIdAndDelete(id)
-        .then((foods) => {
-            // returns the deleted food
-            res.status(200).json({msg: "Following food has been deleted", 
-                food: foods})
-        })
+            .then((foods) => {
+                // returns the deleted food
+                res.status(200).json({
+                    msg: "Following food has been deleted",
+                    food: foods
+                })
+            })
     } catch (error) {
         // logs and returns status 500 if error 
         // food not deleted or couldn't be found
         console.log(error)
-        res.status(500).json({msg: "unable to delete food"})
+        res.status(500).json({ msg: "unable to delete food" })
     }
 }
 
@@ -135,20 +155,17 @@ export const updateFood = async (req, res) => {
         // gets id from params
         const id = req.params.id
 
-        // sets new food with body  
-        const updatedFood = req.body
-
         // updates the food in database
-        await Food.findOneAndUpdate({_id: id}, updatedFood, {new: true})
-        .then((updatedFood) => {
-            // logs and return the updated food
-            console.log(updatedFood)
-            res.status(200).json({msg: "food updated", food: updatedFood})
-        })
+        await Food.findOneAndUpdate({ _id: id }, req.body, { new: true })
+            .then((updatedFood) => {
+                // logs and return the updated food
+                console.log(updatedFood)
+                res.status(200).json({ msg: "food updated", food: updatedFood })
+            })
     } catch (error) {
         // logs and returns status 500 if error 
         // food couldn't be found or couldn't be updated
         console.log(error)
-        res.status(500).json({msg: "unable to update food"})
+        res.status(500).json({ msg: "unable to update food" })
     }
 }
